+++ src/icons/filters-icon.tsx
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2011-2020 Genestack Limited
+ * All Rights Reserved
+ * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF GENESTACK LIMITED
+ * The copyright notice above does not evidence any
+ * actual or intended publication of such source code.
+ */
+
+import * as React from 'react';
+
+import {createIcon} from '../components/icon';
+
+/** Hopper icon */
+export const FiltersIcon = createIcon(
+    <svg viewBox="0 0 20 20">
+        <path d="M3 4v1l6 6v6l2-1v-5l6-6V4H3z" className="gs-svg-icon-fill" />
+    </svg>
+);